Changed file system width and height to use uints instead of ints (remaking of puzzles may be requried)

+++ SilverBotAndGuy/LevelCreator/Form1.cs
@@ -19,11 +19,11 @@ namespace LevelCreator
             InitializeComponent();
         }
 
-        byte[,] GetBytes (out int width, out int height)
+        byte[,] GetBytes (out uint width, out uint height)
         {
             string[] split = textBox1.Text.Replace("\r", "").Split('\n');
-            width = split.OrderByDescending(o=>o.Length).ToArray()[0].Length;
-            height = split.Length;
+            width = (uint)(split.OrderByDescending(o=>o.Length).ToArray()[0].Length);
+            height = (uint)(split.Length);
             byte[,] blocks = new byte[width, height];
             int y = (int)(0);
             int x = (int)(0);
@@ -73,8 +73,8 @@ namespace LevelCreator
                 if (fileStream != null)
                 {
                     BinaryWriter writer = new BinaryWriter(fileStream);
-                    int width;
-                    int height;
+                    uint width;
+                    uint height;
                     byte[,] bytes = GetBytes(out width, out height);
                     if (bytes == null)
                         return;
